+++ Sources/Client/DialogBoxManager.h
@@ -21,11 +21,11 @@ public:
 	IDialogBox* get_dialog_box(int box_id) const;
 	void update_dialog_boxs();
 	void draw_dialog_boxs(short mouse_x, short mouse_y, short z, char lb);
-	void enable_dialog_box(int box_id, int type, int v1, int v2, char* string = nullptr);
-	void enable_dialog_box(DialogBoxId::Type id, int type, int v1, int v2, char* string = nullptr);
+	void enable_dialog_box(int box_id, int type, int64_t v1, int v2, char* string = nullptr);
+	void enable_dialog_box(DialogBoxId::Type id, int type, int64_t v1, int v2, char* string = nullptr);
 	void disable_dialog_box(int box_id);
 	void disable_dialog_box(DialogBoxId::Type id);
-	void toggle_dialog_box(DialogBoxId::Type id, int type = 0, int v1 = 0, int v2 = 0, char* string = nullptr);
+	void toggle_dialog_box(DialogBoxId::Type id, int type = 0, int64_t v1 = 0, int v2 = 0, char* string = nullptr);
 	int get_top_dialog_box_index() const;
 
 	void draw_all(short mouse_x, short mouse_y, short z, char lb);
@@ -42,9 +42,9 @@ public:
 	// Right-click to close dialogs
 	// Returns: true if a dialog was under the mouse (and potentially closed)
 	bool handle_right_click(short mouse_x, short mouse_y, uint32_t time);
-	void enable(DialogBoxId::Type id, int type, int v1, int v2, char* string = nullptr);
+	void enable(DialogBoxId::Type id, int type, int64_t v1, int v2, char* string = nullptr);
 	void disable(DialogBoxId::Type id);
-	void toggle(DialogBoxId::Type id, int type = 0, int v1 = 0, int v2 = 0, char* string = nullptr);
+	void toggle(DialogBoxId::Type id, int type = 0, int64_t v1 = 0, int v2 = 0, char* string = nullptr);
 	int get_top_id() const;
 	bool is_enabled(DialogBoxId::Type id) const;
 	bool is_enabled(int box_id) const;
