+++ Sources/Client/DialogBox_SellList.cpp
@@ -273,7 +273,7 @@ bool DialogBox_SellList::on_item_drop(short mouse_x, short mouse_y)
 		dropInfo.m_v3 = 1001;
 		dropInfo.m_v4 = item_id;
 		std::memset(dropInfo.m_str, 0, sizeof(dropInfo.m_str));
-		m_game->m_dialog_box_manager.enable_dialog_box(DialogBoxId::ItemDropExternal, item_id, m_game->m_item_list[item_id]->m_count, 0);
+		m_game->m_dialog_box_manager.enable_dialog_box(DialogBoxId::ItemDropExternal, item_id, static_cast<int64_t>(m_game->m_item_list[item_id]->m_count), 0);
 		m_game->m_is_item_disabled[item_id] = true;
 	}
 	else
